==> notebooks/01-preprocessing.ipynb <==
## Havoc on the Hill -- A Visualisation of Canadian Legislation Since 1994
By Riley Wheadon  
January 2nd, 2023
---

import pandas as pd
import numpy as np

df = pd.read_json('../data/raw/bills.json')

display(df.head(3))

def printNull(df):
    output = "\n".join([
        "-----------", 
        "{shape}",
        "-----------",
        "{missing}"
    ]).format(
        shape = df.shape, 
        missing = (df
            .isna()
            .sum()
            .to_string()
        )
    )
    
    print(output)

printNull(df)

As shown above, the dataset has a staggering **135** columns. However, multiple columns are missing all 6761 rows of data. We'll start by removing these columns. Additionally, some columns are redundant (i.e. duplicate information in French).

---

df = (df
    .drop(list(df.filter(regex = 'En$|Fr$')), axis = 1)
    .dropna(axis = 1, how = 'all')
    .loc[:, (df != 0).any(axis = 0)]
)

printNull(df)

It appears that the data contains groups of columns with binary data (i.e. PassedFirstChamberFirstReading and similar columns). These groups can be replaced with a single categorical column to improve clarity.

# Aggregate binary 'reading passed' columns into a single column
reading_columns = [col for col in df.columns if col[-7:] == "Reading"]
df["ReadingsPassed"]  = df[reading_columns].astype(int).sum(1)

# Aggregate IsHouseBill and IsSenateBill
df["BillOrigin"] = df[['IsHouseBill', 'IsSenateBill']].idxmax(1)

# Unify the title columns, with a preference for short titles where possible
df['ShortTitle'] = np.where(df['ShortTitle'] == '', df['LongTitle'], df['ShortTitle'])

df = df.drop(columns = reading_columns + ['IsHouseBill', 'IsSenateBill', 'LongTitle'])

with pd.option_context('display.max_columns', None): 
    display(df.head(3))

df = df.drop(
    columns = [
        'LatestCompletedMajorStageId', 'LatestCompletedMajorStageName', 'LatestCompletedMajorStageChamberOrganizationId',
        'LatestBillEventDateTime', 'OriginatingChamberOrganizationId', 'DidReinstateFromPreviousSession', 'BillStages', 
        'StatusName', 'SponsorPersonName'
    ]
)

df = df.rename(
    columns = {
        'NumberCode': 'Code',
        'ShortTitle': 'Title',
        'LatestCompletedMajorStageNameWithChamberSuffix': 'LatestStageName',
        'BillDocumentTypeName': 'BillType',
        'IsSessionOngoing': 'Ongoing'
    }
)

with pd.option_context('display.max_columns', None): 
    display(df.head(3))

# Get DateTime columns
datetime_columns = [col for col in df.columns if col[-8:] == "DateTime"]
df[datetime_columns] = df[datetime_columns].apply(lambda x: pd.to_datetime(x.str[:10]))

# Create new columns and calculate the amount of time the bill was debated
df["FirstStageDate"] = df[datetime_columns].min(axis = 1)
df["LastStageDate"] = df[datetime_columns].max(axis = 1)
df["TimeAlive"] = df["LastStageDate"] - df["FirstStageDate"]

# Drop original DateTime columns
df = df.drop(columns = datetime_columns)

# Replace Id column with a commonly used combination of parliament number, sesion number, and code
df["Id"] = df["ParliamentNumber"].astype(str) + "-" + df["SessionNumber"].astype(str) + "/" +  df["Code"] 

# Print remaining columns and check for null values
printNull(df)
display(df.head(3))

The data has been reduced down to a reasonable size, but there are still some issues. The PersonName column is filled with whitespace and there is no information on political affiliations of the various bills. Webscraping with beautifulsoup4 will be used to get additional data.

--- 

df.to_csv('../data/processed/bills_processed.csv', index = False)
