==> notebooks/03-cleaning.ipynb <==
## Section 1: Aggregating Bill Data

This section will aggregate data from three sources:
- Sponsor information webscraped from the LEGISinfo database (https://www.parl.ca/legisinfo/en/bill/) in 'raw/sponsors.csv'
- Information on the members of parliament from https://www.ourcommons.ca/members/ in 'raw/mp.csv'
- Cleaned bill data from https://www.parl.ca/legisinfo/en/bills?parlsession=all in 'processed/bills_processed.csv'

# Import Python libraries
import pandas as pd
import numpy as np

# Read data files
sponsor_data = pd.read_csv('../data/raw/sponsors.csv')
mp_data = pd.read_csv('../data/raw/mp.csv')
bill_data = pd.read_csv('../data/processed/bills_processed.csv')

# Create helper function to print missing values by column
def printNull(df):
    output = "\n".join([
        "-----------", 
        "{shape}",
        "-----------",
        "{missing}"
    ]).format(
        shape = df.shape, 
        missing = df.isna().sum().to_string()
    )
    print(output)

### Part 1a: Reformatting Sponsor Data

# Split bill_info objects into attributes
sponsor_data[['Id', 'Name', 'Title', 'Constituency']] = sponsor_data['SponsorInfo'].str.split('\n', 3, expand=True)
sponsor_data.drop(columns = ['SponsorInfo'], inplace = True)

for col in sponsor_data.columns:
    sponsor_data[col] = sponsor_data[col].apply(lambda x: x.split(':')[1].strip())
    
# Reformat names column to remove middle initials
def removeInitials(text):
    names = text.split(' ')
    names = [x for x in names if '.' not in x]
    return ' '.join(names)
        
sponsor_data['Name'] = sponsor_data['Name'].apply(lambda x: removeInitials(x))

# Rename columns to specify bill sponsorship
sponsor_data.rename(
    columns = {
        'Name': 'SponsorName',
        'Title': 'SponsorTitle'
    }, 
    inplace = True
)

printNull(sponsor_data)

### Part 1b: Reformatting MP Data

# Merge first and last name columns
mp_data['Name'] = mp_data['First Name'] + ' ' + mp_data['Last Name']

# Selecting only the political affiliation and name columns
mp_data = mp_data[['Political Affiliation', 'Name']]

printNull(mp_data)

### Part 1c: Merging and Exporting

# Merging bill and mps
data = pd.merge(sponsor_data, mp_data, left_on = ['SponsorName'], right_on = ['Name'], how = 'left')

# Get data overview
printNull(data)
display(data.head(3))

# Merge data and bill
data = pd.merge(data, bill_data, on = ['Id'], how = 'left')

# Drop redundant name column
data.drop(columns = 'Name', inplace = True)

# Get data overview
printNull(data)
display(data.head(3))

data.to_csv('../data/cleaned/bill_data.csv', index = False)

## Section 2: Aggregating Parliament Session Data

This section aggregates data from two sources:
- Parliament information from https://en.wikipedia.org/wiki/List_of_Canadian_federal_parliaments in 'raw/parliaments.csv'
- Parliament session information from https://lop.parl.ca/sites/ParlInfo/default/en_CA/Parliament/parliamentsSessions in 'raw/sessions.csv'

### Part 2a: Reformatting Wikipedia Data

parliaments = pd.read_csv('../data/raw/parliaments.csv')

# Cleaning up the data (loaded direct from Wikipedia, really messy)

parliaments.drop(columns = ['Diagram'], inplace = True)
parliaments.drop(list(range(0, 38)) + [39, 41], axis = 0, inplace = True)

parliaments.columns = [
    'ParliamentNumber', 
    'Duration',
    'GoverningParty',
    'Seats',
    'OfficialOpposition',
    'ThirdParties'
]

# Getting ParliamentNumber

parliaments['ParliamentNumber'] = parliaments['ParliamentNumber'].apply(lambda x: int(x[:2]))

# Split GoverningParty column

parliaments[['Party', 'PrimeMinister']] = parliaments['GoverningParty'].str.split('\n', 1, expand = True)
parliaments['PrimeMinister'] = parliaments['PrimeMinister'].apply(lambda x: x.split('\n')[0].split(' ' , 1)[1].split(u'\u2014')[0])
parliaments.drop(columns = 'GoverningParty', inplace = True)

# Create Minority Column

parliaments['Minority'] = parliaments['Seats'].str.contains('minority')

# Drop unnecessary columns

parliaments.drop(columns = ['Seats', 'Duration', 'OfficialOpposition', 'ThirdParties'], inplace = True)

# Check data

display(parliaments.head())

## Part 2b: Reformatting Parliament Session Data

sessions = pd.read_csv('../data/raw/sessions.csv')

# Strip parliament numbers from session column
parliament_numbers = []
def getParliamentNumber(text):
    if 'Parliament:' in text:
        text = text.split(' ', 1)[-1]
        parliament_numbers.append(text)
        
    else:
        text = parliament_numbers[-1]
        
    return text

sessions['ParliamentNumber'] = sessions['Session'].apply(lambda x: getParliamentNumber(x))

# Remove session columns with NaN values

sessions.dropna(how = 'any', subset = ['Start Date'], inplace = True)

# Rename columns

sessions.rename(columns = {
    'Session': 'SessionNumber',
    'Start Date': 'StartDate',
    'End Date': 'EndDate',
    'Senate Sitings': 'SenateSittings',
    'House of Commons Sittings': 'HouseSittings'
}, inplace = True)

# Change datatype of ParliamentNumber and SessionNumber

sessions.loc[:, ['ParliamentNumber', 'SessionNumber']] = sessions.loc[:, ['ParliamentNumber', 'SessionNumber']].astype(int)

# Reverse data, and filter by ParliamentNumber

sessions = sessions.loc[sessions['ParliamentNumber'].astype(int) > 34, :]
sessions = sessions.iloc[::-1]

# Change datatype of Senate and House of Commons Sittings

sessions.iloc[:, -2:] = sessions.iloc[:, -2:].astype(int)

# Check data
print(sessions.dtypes)
display(sessions.head())

## Part 2c: Merging and Exporting

parliaments = pd.merge(parliaments, sessions, on = ['ParliamentNumber'], how = 'right')
parliaments.to_csv('../data/cleaned/parliament_data.csv', index = False)
